+++ datapipe/types.py
@@ -10,10 +10,10 @@ DataSchema = List[Column]
 MetaSchema = List[Column]
 
 # Dataframe with columns (<index_cols ...>)
-IndexDF = NewType('IndexDF', pd.DataFrame)
+IndexDF = NewType("IndexDF", pd.DataFrame)
 
 # Dataframe with columns (<index_cols ...>, hash, create_ts, update_ts, process_ts, delete_ts)
-MetadataDF = NewType('MetadataDF', pd.DataFrame)
+MetadataDF = NewType("MetadataDF", pd.DataFrame)
 
 # Dataframe with columns (<index_cols ...>, <data_cols ...>)
 # DataDF = NewType('DataDF', pd.DataFrame)
@@ -24,7 +24,9 @@ TAnyDF = TypeVar("TAnyDF", pd.DataFrame, IndexDF, MetadataDF)
 
 @dataclass
 class ChangeList:
-    changes: Dict[str, IndexDF] = field(default_factory=lambda: cast(Dict[str, IndexDF], {}))
+    changes: Dict[str, IndexDF] = field(
+        default_factory=lambda: cast(Dict[str, IndexDF], {})
+    )
 
     def append(self, table_name: str, idx: IndexDF) -> None:
         if table_name in self.changes:
@@ -34,7 +36,9 @@ class ChangeList:
             if self_cols != other_cols:
                 raise ValueError(f"Different IndexDF for table {table_name}")
 
-            self.changes[table_name] = cast(IndexDF, pd.concat([self.changes[table_name], idx], axis='index'))
+            self.changes[table_name] = cast(
+                IndexDF, pd.concat([self.changes[table_name], idx], axis="index")
+            )
         else:
             self.changes[table_name] = idx
 
@@ -62,7 +66,7 @@ def meta_to_index(meta_df: MetadataDF, primary_keys: List[str]) -> IndexDF:
 
 
 def index_difference(idx1_df: IndexDF, idx2_df: IndexDF) -> IndexDF:
-    assert(list(idx1_df.columns) == list(idx2_df.columns))
+    assert list(idx1_df.columns) == list(idx2_df.columns)
     cols = idx1_df.columns.to_list()
 
     idx1_idx = idx1_df.set_index(cols).index
@@ -72,7 +76,9 @@ def index_difference(idx1_df: IndexDF, idx2_df: IndexDF) -> IndexDF:
 
 
 def index_intersection(idx1_df: IndexDF, idx2_df: IndexDF) -> IndexDF:
-    assert(sorted(list(idx1_df.columns)) == sorted(list(idx2_df.columns)))  # type: ignore
+    assert sorted(list(idx1_df.columns.tolist())) == sorted(
+        list(idx2_df.columns.tolist())
+    )
     cols = idx1_df.columns.to_list()
 
     idx1_idx = idx1_df.set_index(cols).index